cellemetry_agent/requirements.txt
asyncio==4.0.0
python-dotenv>=1.1.1
fastapi==0.116.1
requests==2.32.5
google-genai==1.54.0
google-adk==1.20.0
matplotlib==3.8.3
numpy==1.24.0
Pillow==12.0.0
protobuf>=5.0
pydantic==2.12.5
python-dotenv==1.2.1
scipy==1.15.3
scikit-image==0.25.2
torch==2.2.0
torchvision==0.17.0
transformers==5.0.0.rc0
uvicorn==0.38.0
xlsxwriter==3.2.9
google-cloud-storage>=3.7.0
